requirements.txt
Django==5.0.6
djangorestframework==3.15.2
dj-database-url==2.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.32.3
